Moved components css to component directory
