+++ src/components/styles/Panel.module.css
